refactor: extract dependency instrospector to external function
In order to improve the maintenance and readability of the code, the introspection of dependencies that was repeated for all tools was taken to an external function.

ref: 10
