+++ templates/github/python/lint.yaml
@@ -12,12 +12,12 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install <%= it.pythonPackageManager.name %>
-          <%= it.pythonPackageManager.dependenciesCommand %>
+          pip install <%= it.packageManager.name %>
+          <%= it.packageManager.dependenciesCommand %>
 
-        <%_ if (!it.flake8Linter.hasFlake8) { %>
+        <%_ if (!it.linters.flake8.hasFlake8) { %>
       - name: Install Flake8
         run: pip install flake8
       <%_ } -%>
 
-      - run: <%= it.pythonPackageManager.runCommand %> flake8
+      - run: <%= it.packageManager.runCommand %> flake8
