+++ templates/github/python/test.yaml
@@ -15,12 +15,12 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install <%= it.pythonPackageManager.name %>
-          <%= it.pythonPackageManager.dependenciesCommand %>
+          pip install <%= it.packageManager.name %>
+          <%= it.packageManager.dependenciesCommand %>
 
-        <%_ if (!it.pytestTester.hasPytest) { %>
+        <%_ if (!it.testers.pytest.hasPytest) { %>
       - name: Install Pytest
         run: pip install pytest
       <%_ } -%>
 
-      - run: <%= it.pythonPackageManager.runCommand %> pytest
+      - run: <%= it.packageManager.runCommand %> pytest
